Split interfaces, remove ignoreWorkingTree left-over
